backend/services/realtime/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
aiokafka==0.11.0
pydantic==2.7.1
python-jose[cryptography]==3.3.0
